Feature_extraction_/requirements.txt
astropy==5.2.2
byol_pytorch==0.8.1
hdbscan==0.8.28
ipython==8.12.3
kornia==0.7.2
matplotlib==3.3.4
numpy==1.24.4
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
pacmap==0.7.3
pandas==1.2.3
Pillow==10.3.0
preprocessing==0.1.13
scikit_learn==1.1.2
scipy==1.8.1
torch==1.12.1+cpu
torchvision==0.13.1+cpu
umap_learn==0.5.2
utils==1.0.2
